+++ DataPersistence/src/main/java/sep3/project/daos/interfaces/LikePersistence.java
@@ -10,9 +10,9 @@ import java.sql.SQLException;
 @Component
 public interface LikePersistence {
 
-    ResponseLikePost LikePost(int postId, long userId) throws SQLException;
-    void UnLikePost(int postId, long userId) throws SQLException;
-    ResponseIsPostLiked IsPostLiked(int postId, long userId) throws SQLException;
-    ResponseCountLikes CountLikes(int postId) throws SQLException;
-    void DeleteLike(int postId) throws SQLException;
+    ResponseLikePost LikePost(long postId, long userId) throws SQLException;
+    void UnLikePost(long postId, long userId) throws SQLException;
+    ResponseIsPostLiked IsPostLiked(long postId, long userId) throws SQLException;
+    ResponseCountLikes CountLikes(long postId) throws SQLException;
+    void DeleteLike(long postId) throws SQLException;
 }
